fix: Android Float type mismatch, Gradle wrapper, iOS build order

+++ android/app/src/main/java/com/example/wearablecollector/viewmodels/FeedViewModel.kt
@@ -30,7 +30,7 @@ class FeedViewModel(private val repository: SocialRepository) : ViewModel() {
                 content = content,
                 timestamp = System.currentTimeMillis(),
                 isPublic = isPublic,
-                stressScore = if (aiStressScore != -1) aiStressScore else null
+                stressScore = if (aiStressScore != -1) aiStressScore.toFloat() else null
             )
             repository.addPost(newPost)
         }
